altered OrderDetails model: Price type from int to double

+++ ProjectWebMVC/Migrations/20230831012341_addOrderHeaderAndDetailsToDb.cs
@@ -55,7 +55,7 @@ namespace ProjectWeb.Migrations
                     OrderHeaderId = table.Column<int>(type: "int", nullable: false),
                     ProductId = table.Column<int>(type: "int", nullable: false),
                     Count = table.Column<int>(type: "int", nullable: false),
-                    Price = table.Column<int>(type: "int", nullable: false)
+                    Price = table.Column<double>(type: "float", nullable: false)
                 },
                 constraints: table =>
                 {
